Add files via upload
The monster!

My boss requested a form that dynamically created itself based upon user input. It was a solo project, so I did the best I could to create his vision.

I never had time to go back and trim this for efficiency or condense it, my apologies for its lack of brevity if you are trying to sort through it.